decoupled engine from gui
